+++ rethinkdb-net/Interfaces/IBaseIndex.cs
@@ -1,0 +1,13 @@
+using System;
+using System.ComponentModel;
+using System.Linq.Expressions;
+
+namespace RethinkDb
+{
+    [ImmutableObject(true)]
+    public interface IBaseIndex<TRecord, TIndex>
+    {
+        ITableQuery<TRecord> Table { get; }
+        string Name { get; }
+    }
+}